+++ src/main/java/usay/app/global/exception/BusinessException.java
@@ -1,7 +1,7 @@
 package usay.app.global.exception;
 
 import lombok.Getter;
-import usay.app.domain.dto.ErrorCode;
+import usay.app.domain.ErrorCode;
 
 @Getter
 public abstract class BusinessException extends RuntimeException {
